% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{export_classifications}
\alias{export_classifications}
\title{Convert a .cla file to a flat table in CSV format}
\usage{
export_classifications(filename, dest)
}
\arguments{
\item{filename}{the name of the file}

\item{dest}{the name of the destination file, by default '<filename.cla>.csv'}
}
\value{
the data.frame invisibly
}
\description{
Convert a .cla file to a flat table in CSV format
}

