% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FlowCam.R
\name{FlowCam}
\alias{FlowCam}
\title{Create an instance of a FlowCamRefClass object}
\usage{
FlowCam(path)
}
\arguments{
\item{path}{character directory name (required, must exist)}
}
\value{
a FlowCamRefClass object
}
\description{
Create an instance of a FlowCamRefClass object
}

