% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{flowcam_read_data}
\alias{flowcam_read_data}
\title{Reads a FlowCam data file, possibly appending post-processed data}
\usage{
flowcam_read_data(X)
}
\arguments{
\item{X}{a FlowCamRefClass}
}
\value{
a data.frame with post-processed columns prepended with "PP"
}
\description{
Reads a FlowCam data file, possibly appending post-processed data
}

