% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FlowCamGroup.R
\name{FlowCamGroup}
\alias{FlowCamGroup}
\title{Read one or more FlowCam datasets}
\usage{
FlowCamGroup(dirlist)
}
\arguments{
\item{dirlist}{character vector of one or more paths to FlowCam datasets}
}
\value{
a list of FlowCam objects in a FlowCamGroup class object
}
\description{
Read one or more FlowCam datasets
}

