% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowcamr-package.R
\docType{package}
\name{flowcamr-package}
\alias{flowcam}
\alias{flowcamr}
\alias{flowcamr-package}
\title{flowcamr-package}
\description{
Tools for working with FlowCam data.
}
\seealso{
\url{http://www.fluidimaging.com/}
}

