% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FlowCam.R
\docType{class}
\name{FlowCamRefClass-class}
\alias{FlowCamRefClass}
\alias{FlowCamRefClass-class}
\title{Reference class for FlowCam data}
\description{
Reference class for FlowCam data
}
\section{Fields}{

\describe{
\item{\code{path}}{the path to the dataset}

\item{\code{name}}{the basename of path}

\item{\code{filelist}}{the listing of relative filepaths}

\item{\code{ctx}}{the context ConfiguratorRefClass}

\item{\code{data}}{a data.frame of .csv exported data}
}}

