% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FlowCam.R
\name{FlowCamRefClass_get_filename}
\alias{FlowCamRefClass_get_filename}
\title{Retrieve a filename by keyword}
\arguments{
\item{what}{the keyword describing the file}

\item{test}{logical if TRUE test for existence}
}
\value{
the fully qualified filename(s)
}
\description{
Retrieve a filename by keyword
}

