% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowCam.R
\name{FlowCamRefClass_which_labeled}
\alias{FlowCamRefClass_which_labeled}
\title{Find the indices of particles labeled as identified}
\arguments{
\item{label}{one or more labels to search for}

\item{column}{the value of the user label column}
}
\value{
zero or more indices of labeled items
}
\description{
Find the indices of particles labeled as identified
}

