% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_metadata.R
\name{get_flowcam_meta}
\alias{get_flowcam_meta}
\title{Harvest FlowCam 'Run Summaries' meta-data from from one or more files}
\usage{
get_flowcam_meta(x, output = NULL)
}
\description{
Harvests meta-data from one or more FlowCam 'Run Summaries' file. Typically,
this won't be called by the user directly. Instead, \code{get_flowcam_meta()} sould be used
instead.
@param x Path to a single FlowCam 'Run Summaries' file or a directory containing multiple
such files.
@param output Path to output .csv file. Defaults to NULL, which will not write a file.
@return A data frame summarizing the metadata from the supplied file(s), optionally also
writing a .csv file.
}
