% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_metadata.R
\name{harvest_fc_meta}
\alias{harvest_fc_meta}
\title{Harvest FlowCam 'Run Summaries' meta-data from a single file}
\usage{
harvest_fc_meta(x)
}
\arguments{
\item{x}{A character vector, resulting from \code{readLines()}.}
}
\value{
A data frame containing harvested meta-data.
}
\description{
Harvests meta-data from a FlowCam 'Run Summaries' file. Typically,
this won't be called by the user directly. Instead, \code{get_flowcam_meta()} should be used
instead.
}
\author{
William Petry \href{mailto:wpetry@ncsu.edu}{wpetry@ncsu.edu}
}
