% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_barcodes.R
\name{mk_barcodes}
\alias{mk_barcodes}
\title{Make barcodes for protist growth assays}
\usage{
mk_barcodes(
  file = "barcodes.pdf",
  experiment = "Pairwise-Assay",
  resident,
  invader = NULL,
  reps = 6,
  byrow = TRUE,
  skip_row = 0L,
  skip_col = 0L,
  sep = "-",
  rm_experiment = TRUE,
  template = "SL582"
)
}
\arguments{
\item{file}{Destination path where PDF of barcodes will be saved.}

\item{experiment}{Experiment name, used as the prefix for the barcode text.}

\item{resident}{Two-letter species code for the resident species.}

\item{invader}{Two-letter species code for the invader species. Use NULL if trial is for a single species.}

\item{reps}{A numeric vector or replicate ids. If a single number, the function will produce a sequence from 1 through the supplied value.}

\item{byrow}{Logical. Whether to fill the page row-wise (TRUE, default) or column-wise (FALSE).}

\item{skip_row}{Number of rows to skip. Useful when a label sheet is partially used.}

\item{skip_col}{Number of columns to skip. Useful when a label sheet is partially used.}

\item{sep}{Separator character used between barcode text chunks.}

\item{rm_experiment}{Logical. Toggles on/off the suppression of 'experiment' in the human-readable text next to the barcode. Does not affect the text embedded in the barcode.}

\item{template}{Template used to format PDF pages. See Details for available options.}
}
\value{
Writes a PDF file to the supplied path.
}
\description{
Make barcodes for protist growth assays
}
\details{
The 'template' option allows users to specify a layout and spacing that corresponds to
the sticky labels on which barcodes are printed. Currently supported labels are:
\itemize{
\item "SL582" (SheetLabels.com) 0.9375"x0.5" labels on 8.5"x11" pages
}
}
\examples{
mk_barcodes(file = "ABxCD_barcodes.pdf", experiment = "competition", resident = "AB", invader = "CD", reps = 3:6)

}
