% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowCamGroup.R
\name{print.FlowCamGroup}
\alias{print.FlowCamGroup}
\title{Pretty print a FlowCamGroup}
\usage{
\method{print}{FlowCamGroup}(x)
}
\arguments{
\item{x}{a FlowCamGroup object}
}
\description{
Pretty print a FlowCamGroup
}

