% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{read_classification_example}
\alias{read_classification_example}
\title{Read an example classification file}
\usage{
read_classification_example(filename = system.file("extdata",
  "201-045311.cla", package = "flowcamr"), ...)
}
\arguments{
\item{filename}{the name of the file to read}

\item{...}{further arguments for \code{read_classification()}}
}
\description{
Read an example classification file
}

