% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{read_classifications}
\alias{read_classifications}
\title{Read a classification table from a .cla file}
\usage{
read_classifications(filename, form = c("data.frame", "FlowCam_class")[1])
}
\arguments{
\item{filename}{the name of the file}

\item{form}{character, desired output - either 'data.frame' or 'FlowCam_class'}
}
\value{
either a FlowCam_class object or a data.frame
}
\description{
Produces a \code{FlowCam_class} object whihc may be converted to \code{data.frame}.
The \code{FlowCam_class} is a list of the following...
\itemize{
   \item version character, the .cla version
   \item n_classifications numeric, the number of classifications
   \item classifications, a list of zero or more \code{klassification} objects
}
}

