% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{read_context}
\alias{read_context}
\title{Read a context file into a Configurator class}
\usage{
read_context(filename, ...)
}
\arguments{
\item{filename}{the name of the context file}

\item{...}{other arguments for \code{Configurator()}}
}
\value{
A ConfiguratorRefClass object
}
\description{
Read a context file into a Configurator class
}
\seealso{
\url{https://github.com/BigelowLab/configurator}
}

