% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{scan_class}
\alias{scan_class}
\title{Scans a class block which include particle ID}
\usage{
scan_class(txt, iline)
}
\arguments{
\item{txt}{the block of text to scan}

\item{iline}{the starting line}
}
\value{
a list of klass and iline
}
\description{
Produces a \code{klass} object which is a list of 
\itemize{
   \item name character, the name of the class (user label)
   \item andor character, either \code{and} or \code{or}
   \item n_filters numeric, the number of filters
   \item filters character, zero or more filter names
   \item n_particles numeric,the number of particles labels in this class
   \item particles character, the particle IDs
}
}

