% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classifier.R
\name{scan_classification}
\alias{scan_classification}
\title{Scans a classification block which is comprised of zero or more classes bundled
into groups of 'klassifcation'}
\usage{
scan_classification(txt, X, iline)
}
\arguments{
\item{txt}{the input text vector}

\item{X}{FlowCam_class object}

\item{iline}{the starting line}
}
\value{
the updated list of X and iline
}
\description{
The \code{klassification} class is a list of the following
\itemize{
   \item name character, the name of the classification group
   \item n_classes numeric, the number of klass objects
   \item classes list, zero of more klass objects
}
}

