% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{show_context}
\alias{show_context}
\title{Show the contents of a context object}
\usage{
show_context(X)
}
\arguments{
\item{X}{the ConfiguratorRefClass context object}
}
\description{
Show the contents of a context object
}

