% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{transform_missing}
\alias{transform_missing}
\title{Identifies 'missingness' in a data frame by examining the value of the
 test_col.  If a given row's (test_col <= 0) & (test_col != NaN) the row 
 values are changed to NA.}
\usage{
transform_missing(x, test_col = "Count", replace = NA,
  keep_col = c("Label", "Blob", "Count", "UserLabel"))
}
\arguments{
\item{x}{a data.frame}

\item{test_col}{the name of the column to test (!=0 and !=NaN)}

\item{replace}{the value to use for the replacement}

\item{keep_col}{the names of the columns to *not* change}

\item{the}{transformed data frame}
}
\description{
Identifies 'missingness' in a data frame by examining the value of the
 test_col.  If a given row's (test_col <= 0) & (test_col != NaN) the row 
 values are changed to NA.
}

