% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{unpack_classes}
\alias{unpack_classes}
\title{Give a list of klasses, return a data.frame of [class_name, particle_id]}
\usage{
unpack_classes(x)
}
\arguments{
\item{x}{a list of klass objects}
}
\value{
data.frame
}
\description{
Give a list of klasses, return a data.frame of [class_name, particle_id]
}

