% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{unpack_classifications}
\alias{unpack_classifications}
\title{Given a FlowCam_class object, return a data.frame of 
    [class_name, particle_id, classification]}
\usage{
unpack_classifications(X)
}
\arguments{
\item{X}{a FlowCam_class object}
}
\value{
data.frame
}
\description{
Given a FlowCam_class object, return a data.frame of 
    [class_name, particle_id, classification]
}

