% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FlowCamGroup.R
\name{volume_summary}
\alias{volume_summary}
\title{Retrieve the volume summary for each FlowCam object.}
\usage{
volume_summary(x, bind = TRUE)
}
\arguments{
\item{x}{a FlowCamGroup}

\item{bind}{logical, if TRUE then rbind the classification tables}
}
\value{
either a list of data.frames or a singke data.frame
}
\description{
Retrieve the volume summary for each FlowCam object.
}

