% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowCamGroup.R
\name{which_labeled}
\alias{which_labeled}
\title{Retrieve the indices for each object for the label(s) specified.}
\usage{
which_labeled(x, ...)
}
\arguments{
\item{x}{a FlowCamGroup}

\item{...}{further arguments for \code{FlowCam}}
}
\value{
a list of indices in each FlowCam dataset for the labels matching the
   requested labels
}
\description{
Retrieve the indices for each object for the label(s) specified.
}
\seealso{
\code{\link{FlowCam_which_labeled}}
}

