% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowCamGroup.R
\name{volume_summary}
\alias{volume_summary}
\title{Retrieve the volume summary for each FlowCam object.}
\usage{
volume_summary(x, bind = TRUE, save_file = TRUE,
  filename = file.path(attr(x, "summary_path"), "volume_summary.csv"))
}
\arguments{
\item{x}{a FlowCamGroup}

\item{bind}{logical, if TRUE then rbind the classification tables}

\item{save_file}{logical, if TRUE then save the output, ignored if 
bind is FALSE}

\item{filename}{the name of the file to save,}
}
\value{
either a list of data.frames or a single data.frame
\dontrun{
 # get a volume-by-class summary table, save by default
 vs <- volume_summary(XX)

 # get a volume-by-class summary table, save to specified file
 vs <- volume_summary(XX, filename = 'C:/my/path/volume_summary.csv')
}
}
\description{
Retrieve the volume summary for each FlowCam object.
}

